[HYJP-002, 003, 007, 008] Created TableView and Search bar
Created a navigation controller which holds the UITableView on the main page. Set the ViewController to look at the new UITableViewController, and prepped the empty cells with a "Picture" identifier. Created a large title, styled with a custom font (Rubik-Bold and Rubik-regular, both of which have been added to the project directory and to the info.plist file) and custom color. The normal title when you scroll up is also styled. Created a search bar using the default searchController (which already has default tapping and typing functionality).